Minor refactor for maintainability.
